www: message-test funnel — /m/[variant] landing pages + lead survey
Four positioning-test landing pages (drive, wiki, connect, assistant) that
reuse the homepage with overridden hero copy. Every signup CTA on a variant
page anchors to a lead survey; submissions email sam@ tagged with the
variant and the ad's utm params. Pages are noindex.

- Extract page.tsx sections into _components/HomePage.tsx with heroOverride
  and surveyVariant props (page files can't have extra named exports)
- SiteHeader takes optional ctaHref
- VariantSurvey: email, agent-usage, role, use-case checkboxes; success
  state notes private beta and links to app signup
- Shared Postmark sender extracted to _lib/postmark.ts (contact-sales now
  imports it)
