+++ www/app/m/actions.ts
@@ -1,0 +1,71 @@
+"use server";
+
+import { escapeHtml, sendPostmark } from "../_lib/postmark";
+
+const LEADS_EMAIL = "sam@joinstash.ai";
+const FROM_ADDRESS = "Stash <notifications@joinstash.ai>";
+
+export type VariantSurveyState = {
+  status: "idle" | "ok" | "error";
+  message?: string;
+};
+
+export async function submitVariantSurvey(
+  _prev: VariantSurveyState,
+  formData: FormData,
+): Promise<VariantSurveyState> {
+  const email = String(formData.get("email") ?? "").trim();
+  const agentUsage = String(formData.get("agentUsage") ?? "").trim();
+  const roleCompany = String(formData.get("roleCompany") ?? "").trim();
+  const useCases = formData.getAll("useCases").map(String);
+  const otherUseCase = String(formData.get("otherUseCase") ?? "").trim();
+  const variant = String(formData.get("variant") ?? "").trim();
+  const utm = String(formData.get("utm") ?? "").trim();
+
+  if (!email || !email.includes("@")) {
+    return { status: "error", message: "Please enter a valid email address." };
+  }
+
+  const token = process.env.POSTMARK_SERVER_TOKEN;
+  if (!token) {
+    console.error("POSTMARK_SERVER_TOKEN is not set; cannot deliver survey submission", {
+      email,
+      variant,
+    });
+    return {
+      status: "error",
+      message: "Signups are not configured yet. Email sam@joinstash.ai directly.",
+    };
+  }
+
+  const leadHtml = `
+    <h2>Messaging test lead</h2>
+    <p><strong>Variant:</strong> ${escapeHtml(variant) || "—"}</p>
+    <p><strong>Email:</strong> ${escapeHtml(email)}</p>
+    <p><strong>Uses AI coding agents:</strong> ${escapeHtml(agentUsage) || "—"}</p>
+    <p><strong>Role / company:</strong> ${escapeHtml(roleCompany) || "—"}</p>
+    <p><strong>Use cases:</strong> ${useCases.map(escapeHtml).join(", ") || "—"}</p>
+    <p><strong>Other use case:</strong></p>
+    <p>${escapeHtml(otherUseCase).replace(/\n/g, "<br/>") || "—"}</p>
+    <p><strong>UTM:</strong> ${escapeHtml(utm) || "—"}</p>
+  `;
+
+  const leadRes = await sendPostmark(token, {
+    From: FROM_ADDRESS,
+    To: LEADS_EMAIL,
+    ReplyTo: email,
+    Subject: `Messaging test lead — ${email} [${variant || "unknown"}]`,
+    HtmlBody: leadHtml,
+    MessageStream: "outbound",
+  });
+
+  if (!leadRes.ok) {
+    console.error("Postmark survey lead send failed", leadRes.status, await leadRes.text());
+    return {
+      status: "error",
+      message: "We couldn't save your request. Please email sam@joinstash.ai.",
+    };
+  }
+
+  return { status: "ok" };
+}
